backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.4
pytest==7.4.2
pytest-flask==1.2.0
redis==4.6.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
celery==5.3.2
gunicorn==21.2.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
imagehash==4.3.1
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
msal==1.24.0
dropbox==11.36.2